requirements.txt
Flask==3.0.0
Pillow==10.1.0
Werkzeug==3.0.1
requests==2.31.0
openai==1.12.0
openai>=1.55.3,<2.0.0
httpx>=0.27.0,<0.28.0


